requirements.txt
pygooglenews
google-genai
python-dotenv
trafilatura
